Hello BorisPolonsky,
Thanks for the wonderful helm solution so that anyone can deploy Dify in Kubernetes.
That being said I have identified an issue where the repository lacked example values for tolerations and affinity for the Postgres (master & replicas) deployment.
This PR includes an updated values.yaml file with a commented example showcasing how to configure these settings for the Postgres database.
Best
Shomu